requirements.txt
Django
django-cors-headers
six
django-filter
djangorestframework
djangorestframework-jwt
PyJWT
pytz
psycopg2
gunicorn
django-heroku
